+++ src/deploy/hosting/prepare.js
@@ -23,37 +23,21 @@ module.exports = function(context, options) {
     options.config.set("hosting.public", options.public);
   }
 
-  const configs = normalizedHostingConfigs(options);
+  const configs = normalizedHostingConfigs(options, { resolveTargets: true });
   if (configs.length === 0) {
     return Promise.resolve();
   }
 
   context.hosting = {
     deploys: configs.map(function(cfg) {
-      return { config: cfg };
+      return { config: cfg, site: cfg.site };
     }),
   };
 
   const versionCreates = [];
 
   _.each(context.hosting.deploys, function(deploy) {
-    let cfg = deploy.config;
-
-    if (cfg.target) {
-      const matchingTargets = options.rc.requireTarget(options.project, "hosting", cfg.target);
-      if (matchingTargets.length > 1) {
-        throw new FirebaseError(
-          `Hosting target ${clc.bold(cfg.target)} is linked to multiple sites, ` +
-            `but only one is permitted. ` +
-            `To clear, run:\n\n  firebase target:clear hosting ${cfg.target}`
-        );
-      }
-      deploy.site = matchingTargets[0];
-    } else if (cfg.site) {
-      deploy.site = cfg.site;
-    } else {
-      throw new FirebaseError('Must supply either "site" or "target" in each "hosting" config.');
-    }
+    const cfg = deploy.config;
 
     if (!cfg.public) {
       throw new FirebaseError(
@@ -85,15 +69,17 @@ module.exports = function(context, options) {
       }
     }
 
+    const data = {
+      config: convertConfig(cfg),
+      labels: deploymentTool.labels,
+    };
+
     versionCreates.push(
       api
-        .request("POST", "/v1beta1/sites/" + deploy.site + "/versions", {
+        .request("POST", `/v1beta1/sites/${deploy.site}/versions`, {
           origin: api.hostingApiOrigin,
           auth: true,
-          data: {
-            config: convertConfig(cfg),
-            labels: deploymentTool.labels,
-          },
+          data,
         })
         .then(function(result) {
           deploy.version = result.body.name;